Many changes. MessageBlock componnet added. ConfirmationModal added.
